+++ hamp_radar/iquick.py
@@ -6,6 +6,7 @@ import xarray as xr
 from numpy.lib.stride_tricks import as_strided
 
 from geometries import SingleSubBlockGeometry, SingleMainBlockGeometry, MultiMainBlockGeometry
+from decoders import decode_srvi, decode_moment, decode_iq, decode_time
 
 def main_ofs(mainblock):
     # blocktype and blocksize assumed to by 4 bytes long
@@ -209,87 +210,6 @@ def extract_raw_arrays(data, mmbgs: Iterable[MultiMainBlockGeometry]):
                 ),
             )
 
-
-def decode_srvi(rawdata):
-    return {
-        "frm": (
-            ("frame",),
-            rawdata[:, 0:4].view("<u4")[:, 0],
-            {"long_name": "data frame number"},
-        ),
-        "Tm": (("frame",), rawdata[:, 4:8].view("<u4")[:, 0]),
-        "TPow": (
-            ("frame",),
-            rawdata[:, 8:12].view("<f4")[:, 0],
-            {"long_name": "avg transmit power"},
-        ),
-        "NPw": (
-            ("frame", "cocx"),
-            rawdata[:, 12:20].view("<f4"),
-            {"long_name": "noise power pin-mod in save position"},
-        ),
-        "CPw": (
-            ("frame", "cocx"),
-            rawdata[:, 20:28].view("<f4"),
-            {"long_name": "noise power int. source"},
-        ),
-        "PS_Stat": (("frame",), rawdata[:, 28:32].view("<u4")[:, 0]),
-        "RC_Err": (("frame",), rawdata[:, 32:36].view("<u4")[:, 0]),
-        "TR_Err": (("frame",), rawdata[:, 36:40].view("<u4")[:, 0]),
-        "dwSTAT": (("frame",), rawdata[:, 40:44].view("<u4")[:, 0]),
-        "dwGRST": (("frame",), rawdata[:, 44:48].view("<u4")[:, 0]),
-        "AzmPos": (("frame",), rawdata[:, 48:52].view("<f4")[:, 0]),
-        "AzmVel": (("frame",), rawdata[:, 52:56].view("<f4")[:, 0]),
-        "ElvPos": (("frame",), rawdata[:, 56:60].view("<f4")[:, 0]),
-        "ElvVel": (("frame",), rawdata[:, 60:64].view("<f4")[:, 0]),
-        "NorthAngle": (("frame",), rawdata[:, 64:68].view("<f4")[:, 0]),
-        "time_milli": (("frame",), rawdata[:, 68:72].view("<u4")[:, 0]),
-        "PD_DataQuality": (("frame",), rawdata[:, 72:76].view("<u4")[:, 0]),
-        "LO_Frequency": (("frame",), rawdata[:, 76:80].view("<f4")[:, 0]),
-        "DetuneFine": (("frame",), rawdata[:, 80:84].view("<f4")[:, 0]),
-    }
-
-
-def decode_moment(name):
-    def _decode(rawdata):
-        # TODO(ALL): HACK: this arbitrarily reduces the range dimension to 512 to fit with the IQ output
-        return {
-            name: (
-                ("frame", "range", "cocx"),
-                rawdata.view("<f4").reshape(rawdata.shape[0], -1, 2)[:, :512, :],
-            )
-        }
-
-    return _decode
-
-
-def decode_iq(rawdata):
-    # TODO(ALL) HACK: the 256 (for nfft) just appears, it probably should be read from somewhere else in the data
-    return {
-        "FFTD": (
-            ("frame", "range", "cocx", "fft", "iq"),
-            rawdata.view("<i2").reshape(rawdata.shape[0], -1, 2, 256, 2),
-        )
-    }
-
-
-# TODO(ALL): move decoders to seperate file
-"""
-Decoders for IQ data as in Meteorological Ka-Band Cloud Radar MIRA35 Manual,
-section 2.3.3.2 'Embedded chain type 2; Data chain'. Note these decoders are
-specific to the Ka radar currently in operation on HALO.
-(last checked: 13th Septermber 2024).
-"""
-decoders = {
-    b"SRVI": decode_srvi,
-    b"SNRD": decode_moment("SNRD"),
-    b"VELD": decode_moment("VELD"),
-    b"HNED": decode_moment("HNED"),
-    b"RMSD": decode_moment("RMSD"),
-    b"FFTD": decode_iq,  # TODO(ALL) HACK: FFTD may or may not be IQ data. This is configured in PPAR
-}
-
-
 def single_dspparams_data(data):
     """
     Returns raw arrays of data from the first occurrence of DSP parameters
@@ -344,6 +264,20 @@ def read_pds(filename, postprocess=True):
     Returns:
     xarray.Dataset: The IQ data dataset.
     """
+
+    # Decoders for IQ data as in Meteorological Ka-Band Cloud Radar MIRA35 Manual,
+    # section 2.3.3.2 'Embedded chain type 2; Data chain'. Note these decoders are
+    # specific to the Ka radar currently in operation on HALO.
+    # (last checked: 13th Septermber 2024).
+    decoders = {
+        b"SRVI": decode_srvi,
+        b"SNRD": decode_moment("SNRD"),
+        b"VELD": decode_moment("VELD"),
+        b"HNED": decode_moment("HNED"),
+        b"RMSD": decode_moment("RMSD"),
+        b"FFTD": decode_iq,  # TODO(ALL) HACK: FFTD may or may not be IQ data. This is configured in PPAR
+    }
+
     data = np.memmap(filename, mode="r")
     raw_arrays = single_dspparams_data(data)
     ds = xr.Dataset(
@@ -358,32 +292,9 @@ def read_pds(filename, postprocess=True):
         ds = ds.pipe(postprocess_iq)
     return ds
 
-
-def decode_time(ds):
-    """
-    Replaces 'Tm' and 'time_milli' variables in dataset 'ds' with decoded time.
-
-    Parameters:
-    ds (xarray.Dataset): The input dataset containing 'Tm' and 'time_milli'
-                         variables.
-
-    Returns:
-    xarray.Dataset: The dataset with 'Tm' and 'time_milli' variables replaced by
-                    new 'time' variable added.
-    """
-    time = (
-        np.datetime64("1970-01-01")
-        + ds.Tm * np.timedelta64(1000000000, "ns")
-        + ds.time_milli
-        * np.timedelta64(1000, "ns")  # [sic] 'time_milli' is microseconds
-    )
-    return ds.drop_vars(["Tm", "time_milli"]).assign(time=time)
-
-
 def postprocess_iq(ds):
     # TODO(ALL): move to new file
     return ds.pipe(decode_time)
-
 
 def untangle_iqf(
     files,
